requirements.txt
ftfy
pytorch
nltk
tqdm
tiktoken